# task-management-backend/src/main/java/com/company/taskmanagementbackend/auth/AuthResponse.java
package com.company.taskmanagementbackend.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthResponse {

    private String token;

    private String fullName;
}