# task-management-backend/src/main/java/com/company/taskmanagementbackend/project/ProjectResponse.java
package com.company.taskmanagementbackend.project;


import lombok.*;

import java.time.LocalDate;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ProjectResponse {

    private Integer id;
    private String name;
    private String description;
    private LocalDate startDate;
    private LocalDate endDate;
    private String status;
    private String ownerName;
}