# task-management-frontend/src/app/services/models/notification-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface NotificationResponse {
  id?: number;
  message?: string;
  read?: boolean;
  timestamp?: string;
  userName?: string;
}