# task-management-frontend/src/app/services/models/page-response-completed-projects-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CompletedProjectsResponse } from '../models/completed-projects-response';
export interface PageResponseCompletedProjectsResponse {
  content?: Array<CompletedProjectsResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}