# task-management-frontend/src/app/services/models/project-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProjectResponse {
  description?: string;
  endDate?: string;
  id?: number;
  name?: string;
  ownerName?: string;
  startDate?: string;
  status?: string;
}