# task-management-frontend/src/app/services/models/task-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TaskRequest {
  description: string;
  dueDate: string;
  id?: number;
  priority: string;
  taskStatus: string;
  title: string;
}