# task-management-frontend/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification","description":"OpenApi documentation for Task Management App","contact":{"name":"Konrad"}},"servers":[{"url":"http://localhost:8080/api","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/task":{"get":{"tags":["Task"],"operationId":"findAllTasks","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseTaskResponse"}}}}}},"post":{"tags":["Task"],"operationId":"saveTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/task/edit-task/{task-id}":{"post":{"tags":["Task"],"operationId":"editTask","parameters":[{"name":"task-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/task/delete/{task-id}":{"post":{"tags":["Task"],"operationId":"deleteTask","parameters":[{"name":"task-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/task/assign/{project-id}":{"post":{"tags":["Task"],"operationId":"assignTaskToProject","parameters":[{"name":"project-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/project":{"get":{"tags":["Project"],"operationId":"findAllProjects","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseProjectResponse"}}}}}},"post":{"tags":["Project"],"operationId":"saveProject","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/project/edit-project/{project-id}":{"post":{"tags":["Project"],"operationId":"editProject","parameters":[{"name":"project-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/project/delete/{project-id}":{"post":{"tags":["Project"],"operationId":"deleteProject","parameters":[{"name":"project-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/notifications":{"get":{"tags":["notification-controller"],"operationId":"findAllNotifications","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseNotificationResponse"}}}}}},"post":{"tags":["notification-controller"],"operationId":"saveNotification","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/project/completed/{project-id}":{"patch":{"tags":["Project"],"operationId":"markProjectAsCompleted","parameters":[{"name":"project-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/notifications/read/{notification-id}":{"patch":{"tags":["notification-controller"],"operationId":"markNotificationAsRead","parameters":[{"name":"notification-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/task/{task-id}":{"get":{"tags":["Task"],"operationId":"findTaskById","parameters":[{"name":"task-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}}}}},"/project/{project-id}":{"get":{"tags":["Project"],"operationId":"findProjectById","parameters":[{"name":"project-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}}}}},"/project/search":{"get":{"tags":["Project"],"operationId":"searchByNameContainingIgnoreCase","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseProjectResponse"}}}}}}},"/project/owner":{"get":{"tags":["Project"],"operationId":"findAllProjectsByOwner","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseProjectResponse"}}}}}}},"/project/ending":{"get":{"tags":["Project"],"operationId":"getEndingProjects","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseProjectResponse"}}}}}}},"/project/completed":{"get":{"tags":["Project"],"operationId":"findAllCompletedProjects","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseCompletedProjectsResponse"}}}}}}},"/notifications/{notification-id}":{"get":{"tags":["notification-controller"],"operationId":"findNotificationById","parameters":[{"name":"notification-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationResponse"}}}}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirmAccount","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"TaskRequest":{"required":["description","dueDate","priority","taskStatus","title"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"description":{"type":"string"},"taskStatus":{"type":"string"},"priority":{"type":"string"},"dueDate":{"type":"string","format":"date"}}},"ProjectRequest":{"required":["description","endDate","name","startDate","status"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"status":{"type":"string"}}},"NotificationRequest":{"required":["message","timestamp"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"message":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"RegistrationRequest":{"required":["email","firstName","lastName","password"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"fullName":{"type":"string"}}},"PageResponseTaskResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/TaskResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"},"first":{"type":"boolean"}}},"TaskResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"description":{"type":"string"},"taskStatus":{"type":"string"},"priority":{"type":"string"},"dueDate":{"type":"string","format":"date"},"assignedTo":{"type":"string"},"projectName":{"type":"string"}}},"PageResponseProjectResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/ProjectResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"},"first":{"type":"boolean"}}},"ProjectResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"status":{"type":"string"},"ownerName":{"type":"string"}}},"CompletedProjectsResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"endDate":{"type":"string","format":"date"},"ownerName":{"type":"string"},"completed":{"type":"boolean"}}},"PageResponseCompletedProjectsResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/CompletedProjectsResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"},"first":{"type":"boolean"}}},"NotificationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"message":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"read":{"type":"boolean"},"userName":{"type":"string"}}},"PageResponseNotificationResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/NotificationResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"},"first":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}